complete alias methods for case participated model
